[feat] SatisfictionChip 구현
