+++ SMASHING/Global/Coordinators/Coordinator.swift
@@ -31,7 +31,7 @@ extension Coordinator {
         let toast = ToastMessage()
         navigationController.viewControllers.last?.view.addSubview(toast)
         
-        toast.configure(title: notificationType.message, action: { [weak self] in
+        toast.configure(title: notificationType.rawValue, action: { [weak self] in
             guard let self = self else { return }
             
             let transition = CATransition().then {
