+++ SMASHING/Global/Enums/ReviewScore.swift
@@ -5,6 +5,8 @@
 //  Created by 이승준 on 1/13/26.
 //
 
+import UIKit
+
 enum ReviewScore: String, Codable {
     case bad = "별로에요"
     case good = "좋아요"
@@ -14,5 +16,21 @@ enum ReviewScore: String, Codable {
         case bad = "BAD"
         case good = "GOOD"
         case best = "BEST"
+    }
+    
+    var imageLarge: UIImage {
+        switch self {
+        case .bad: return UIImage.icThumbsDownLg
+        case .good: return UIImage.icThumbsUpLg
+        case .best: return UIImage.icThumbsUpDoubleLg
+        }
+    }
+    
+    var imageSmall: UIImage {
+        switch self {
+        case .bad: return UIImage.icThumbsDown
+        case .good: return UIImage.icThumbsUp
+        case .best: return UIImage.icThumbsUpDouble
+        }
     }
 }
