+++ SMASHING/Global/Resource/Colors.xcassets/Review/ic_thumbs_up_lg.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "ic_thumbs_up_lg.pdf",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
