+++ SMASHING/Presentation/Review/MyReviewsView.swift
@@ -10,62 +10,6 @@ import UIKit
 import SnapKit
 import Then
 
-final class SatisfictionChip: BaseUIView {
-    
-    // MARK: - UI Components
-    
-    private let contentStackView = UIStackView().then {
-        $0.axis = .horizontal
-        $0.alignment = .center
-        $0.spacing = 10
-        $0.isUserInteractionEnabled = false
-    }
-    
-    private let imageView = UIImageView().then {
-        $0.contentMode = .scaleAspectFit
-    }
-    
-    private let label = UILabel().then {
-        $0.font = .pretendard(.textSmR)
-        $0.textColor = .Text.primary
-    }
-    
-    // MARK: - Setup Methods
-    
-    init() {
-        // self.review = review // review Tpye 정의 이후,
-        super.init(frame: .zero)
-        setupAttributes()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func setupAttributes() {
-        self.layer.cornerRadius = 20
-        // self.imageView.image = reivew.image
-    }
-
-    override func setUI() {
-        addSubview(contentStackView)
-        [imageView, label].forEach { contentStackView.addArrangedSubview($0) }
-    }
-
-    override func setLayout() {
-        contentStackView.snp.makeConstraints {
-            $0.leading.trailing.equalToSuperview().inset(16)
-            $0.top.bottom.equalToSuperview().inset(10)
-            $0.height.equalTo(24)
-        }
-        
-        imageView.snp.makeConstraints {
-            $0.size.equalTo(24)
-        }
-    }
-
-}
-
 final class MyReviewsView: BaseUIView {
     
     // MARK: - Properties
@@ -122,6 +66,11 @@ final class MyReviewsView: BaseUIView {
         addSubviews(navigationBar, satisfactionLabel, satisfactionStackView,
                     quickReviewLabel, quickReviewStackView,
                     allReviewLabel, reviewCollectionView)
+        
+        [ReviewScore.best, .good, .bad].forEach { review in
+            let chip = SatisfictionChip(review: review, num: Int.random(in: 0...150))
+            satisfactionStackView.addArrangedSubview(chip)
+        }
     }
     
     override func setLayout() {
